Add new Statistics for weight of SCoPs.
This adds a new statistic that holds the 'weight' of a SCoP. A SCoP's
weight is defined as the number of subregions that are contained within
the SCoP.
We consider SCops that are expanded from smaller regions to be more
'impactfull' for the program's performance than small (canonical ones).

Therefore, the more regions we contain inside a single SCoP the better.